Build go-ipfs from git
We now build both js-ipfs and go-ipfs from github master each time.

The support for specific 'commit' tests are currently broken, as
the scripts will attempt to checkout the same commit on both js-ipfs
as well as go-ipfs ... we'll need to switch on the target as well
as pick a default for the other.

I've hardcoded some paths for the go-ipfs binary we build ... that
should be fixed.

I added an ansible role to install an up-to-date version of go.

There are multiple versions of go and ipfs on the system, so it's
a bit hazardous as the wrong version might be used by accident.